Start moving `plugin.defaultHost` to it's own package
