+++ pkg/host/testdata/analyzer-language/pulumi-language-test
