+++ pkg/host/testdata/analyzer-no-config/main.go
