+++ pkg/host/testdata/analyzer-no-config/pulumi-analyzer-policy-test
