+++ pkg/host/testdata/analyzer/main.go
