+++ pkg/host/testdata/analyzer/pulumi-analyzer-policy-test
