+++ pkg/host/testdata/policypack/PulumiPolicy.yaml
