+++ pkg/host/testdata/provider-language/pulumi-language-test.cmd
