+++ pkg/host/testdata/test-plugin-cli-version/PulumiPlugin.yaml
