+++ pkg/host/testdata/test-plugin-exit/PulumiPlugin.yaml
