+++ pkg/host/testdata/test-plugin/PulumiPlugin.yaml
